youtube_uploader/requirements-minimal.txt
google-api-python-client
google-auth-oauthlib
httplib2
pyyaml
